chore: merge monster and character
