+++ app/Models/Characters/Character.php
@@ -6,6 +6,9 @@ use App\Models\Core\Race;
 use App\Models\Core\Size;
 use App\Models\Core\Dice;
 use App\Models\Core\Alignment;
+use App\Models\Core\Skill;
+use App\Models\Items\Item;
+use App\Models\Language\Language;
 use App\Traits\CharacterSkillsTrait;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
@@ -39,6 +42,15 @@ class Character extends Model
         return $this->belongsTo(Alignment::class);
     }
 
+    public function languages()
+    {
+        return $this->belongsToMany(
+            Language::class,
+            'character_languages',
+            'character_id'
+        );
+    }
+
     public function race()
     {
         return $this->belongsTo(Race::class);
@@ -54,9 +66,19 @@ class Character extends Model
         return $this->belongsTo(Size::class);
     }
 
+    public function items()
+    {
+        return $this->belongsToMany(Item::class);
+    }
+
+    public function skills()
+    {
+        return $this->belongsToMany(Skill::class)->withPivot('bonus');
+    }
+
     /** END RELATIONS */
 
-    public function getAbilitiesAttribute()
+    public function getAbilityModifiersAttribute()
     {
         return Arr::only(
             $this->toArray(),
@@ -88,12 +110,6 @@ class Character extends Model
                 ]
             )
         );
-    }
-
-    public function hp_to_string()
-    {
-        $die = "{$this->hp_dice_count}d{$this->hp_dice->sides}";
-        return "$die + $this->base_hp";
     }
 
     public function getAverageHpAttribute()
@@ -138,6 +154,15 @@ class Character extends Model
         return (int) floor(($ability_score - 10) / 2);
     }
 
+    public function getSkillBonus(string $skill_name)
+    {
+        $skill = $this->skills()
+            ->where('code', $skill_name)
+            ->first();
+
+        return $skill ? $skill->pivot->bonus : 0;
+    }
+
     public function getArmorClassAttribute()
     {
         $armor_class = 10;
@@ -156,5 +181,73 @@ class Character extends Model
                 },
                 $armor_class
             );
+    }
+
+    public function getArmorAttribute()
+    {
+        return $this->items()
+            ->whereHas('type', function ($type) {
+                $type->where('code', 'armor');
+            })
+            ->where('item_monster.equipped', true)
+            ->whereDoesntHave('subtype', function ($subtype) {
+                $subtype->where('code', 'shield');
+            })
+            ->first();
+    }
+
+    public function getShieldAttribute()
+    {
+        return $this->items()
+            ->whereHas('type', function ($type) {
+                $type->where('code', 'armor')
+                    ->where('equipped', true);
+            })
+            ->whereHas('subtype', function ($subtype) {
+                $subtype->where('code', 'shield');
+            })
+            ->first();
+    }
+
+    public function getWeaponsAttribute()
+    {
+        return $this->items()
+            ->with(['actions'])
+            ->whereHas('type', function ($type) {
+                $type->where('code', 'weapon');
+            })
+            ->where('item_monster.equipped', true)
+            ->get();
+    }
+
+    public function getProficiencyBonusAttribute()
+    {
+        /**
+         * proficiency bonus starts at 2 and goes up by 1 at level 5 and every
+         * 4 levels thereafter. The formula for proficiency bonus is therefore:
+         *
+         * 1 + lvl / 4 rounded up
+         */
+        return ceil(1 + ($this->challenge->level / 4));
+    }
+
+    public function getArmorModifiers()
+    {
+        return $this->items()
+            ->with('modifiers')
+            ->where('item_monster.equipped', true)
+            ->whereHas('type', function ($type) {
+                $type->where('code', 'armor');
+            })
+            ->whereHas('modifiers', function ($modifiers) {
+                $modifiers->where('type', 'stat')
+                    ->where('code', 'armor_class');
+            })
+            ->get()
+            ->pluck('modifiers')
+            ->collapse()
+            ->filter(function ($modifier) {
+                return $modifier->type === 'stat' && $modifier->code === 'armor_class';
+            });
     }
 }
